requirements.txt
os-sys==2.1.4
jsonlib==1.6.2
tkinter==0.1.0
mido==1.2.10
rtmidi-python==0.6.4
threading==1.3.4
struct==0.4.0
